bot-frontend/pyproject.toml
[tool.poetry]
name = "bot-frontend"
version = "0.1.0"
description = ""
authors = ["infiniteregrets <aroram18@mcmaster.ca>"]
readme = "README.md"
packages = [{include = "bot_frontend"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
